requirements.txt
python-dotenv
pydantic


# For Langchain
langchain==0.3.4
langchain-openai==0.2.3
langchain-community==0.3.3
langchain-chroma
langchain_postgres
chroma-hnswlib==0.7.3
chromadb==0.5.0


# For MCP
fastmcp

# misc
tqdm
pandas
openpyxl
xlsxwriter